PR #84: Move to simpler CrossEntropyLoss
The hidden cells are added in 01 to improve the flow. The notebook
a little text-heavy.

Address review comments and fix formatting

- Fix a few more typos
- Remove <details> tag in Dataset class markdown description. Add
ellipsis
- Rework the CrossEntropyLoss description as an alternative to BCELoss
- Move the Resnet example up a few cells so it flows better with the
network exercise
- Remove hint for the iterate over features exercise
- Remove comment about activation function not being in the network

Make initial changes to move to CrossEntropyLoss

Convert notebook 1 to work with CrossEntropyLoss

Modified link and changed comment to CELoss

Replace BCE with CrossEntropy in exercise

Modify content on Softmax in Task 9.

It's not necessary to use softmax in the notebook as it's
already included in the PyTorch implementation. Add a bit more
explanation about this in the notebook.

Add note on excluding activation from model

Inital fix to CrossEntropy in 03

Fix eye typos

Add executed 03

Update Resnet50 name, add note in task 9

Pylance fixes and run 01 solution